refactor(split-rendering): move render_content / compute_content_layout into orchestration (Phase 5)
Shrink mod.rs from a big impl block into a thin SplitRenderer façade:

- Move render_content (~695 lines), compute_content_layout (~137 lines),
  and build_base_tokens_for_hook into orchestration/mod.rs as pub(crate)
  free functions
- Keep SplitRenderer struct with three pub fn forwarding methods so the
  crate's external call sites (app/render.rs) keep working unchanged
- Scope the tests module's imports locally instead of pulling everything
  in via `use super::*;`

Invariant check passes: no self-contained top-level module
(spans, style, char_style, post_pass, base_tokens, transforms, view_data,
folding, scrollbar, layout, gutter) imports SelectionContext or
DecorationContext. Only files under orchestration/ reference them.

All 37 split_rendering tests pass.
